Move all classes into packages
All classes are currently in default package.

Let's create new packages and move them into their respective
packages, adding imports as necessary.
