+++ src/main/java/astraea/command/Command.java
@@ -1,3 +1,9 @@
+package astraea.command;
+
+import astraea.storage.Storage;
+import astraea.task.TaskList;
+import astraea.ui.AstraeaUI;
+
 public class Command {
     private final CommandType commandType;
     private final String[] args;
