+++ src/main/java/astraea/command/CommandType.java
@@ -1,3 +1,5 @@
+package astraea.command;
+
 public enum CommandType {
     LIST, MARK, UNMARK, TODO, DEADLINE, EVENT, DELETE, EXIT
 }
