+++ src/main/java/astraea/command/DeadlineCommand.java
@@ -1,16 +1,23 @@
+package astraea.command;
+
+import astraea.task.Deadline;
+import astraea.ui.AstraeaUI;
+
+import astraea.storage.Storage;
+import astraea.task.TaskList;
 import java.io.IOException;
 
-public class TodoCommand extends Command {
-    public TodoCommand(CommandType type, String[] args) {
+public class DeadlineCommand extends Command {
+    public DeadlineCommand(CommandType type, String[] args) {
         super(type, args);
     }
 
     @Override
     public void execute(TaskList list, Storage storage, AstraeaUI ui) {
-        Todo task = new Todo(this.getArguments()[0]);
+        Deadline task = Deadline.createDeadline(this.getArguments()[0], this.getArguments()[1]);
         list.add(task);
         String[] message = new String[]{
-                "Much ado about this todo.",
+                "Time's ticking on this deadline. Get to it soon.",
                 "  " + task,
                 "I'm tracking " + list.size() + " of your tasks now."
         };
