+++ src/main/java/astraea/command/DeleteCommand.java
@@ -1,3 +1,10 @@
+package astraea.command;
+
+import astraea.storage.Storage;
+import astraea.task.Task;
+import astraea.task.TaskList;
+import astraea.ui.AstraeaUI;
+
 import java.io.IOException;
 
 public class DeleteCommand extends Command {
