+++ src/main/java/astraea/command/EventCommand.java
@@ -1,16 +1,23 @@
+package astraea.command;
+
+import astraea.storage.Storage;
+import astraea.task.Event;
+import astraea.task.TaskList;
+import astraea.ui.AstraeaUI;
+
 import java.io.IOException;
 
-public class DeadlineCommand extends Command {
-    public DeadlineCommand(CommandType type, String[] args) {
+public class EventCommand extends Command {
+    public EventCommand(CommandType type, String[] args) {
         super(type, args);
     }
 
     @Override
     public void execute(TaskList list, Storage storage, AstraeaUI ui) {
-        Deadline task = Deadline.createDeadline(this.getArguments()[0], this.getArguments()[1]);
+        Event task = Event.createEvent(this.getArguments()[0], this.getArguments()[1], this.getArguments()[2]);
         list.add(task);
         String[] message = new String[]{
-                "Time's ticking on this deadline. Get to it soon.",
+                "A fleeting moment in time to be. Don't miss this.",
                 "  " + task,
                 "I'm tracking " + list.size() + " of your tasks now."
         };
