+++ src/main/java/astraea/command/ExitCommand.java
@@ -1,3 +1,9 @@
+package astraea.command;
+
+import astraea.storage.Storage;
+import astraea.task.TaskList;
+import astraea.ui.AstraeaUI;
+
 public class ExitCommand extends Command {
     public ExitCommand(CommandType type, String[] args) {
         super(type, args);
