+++ src/main/java/astraea/command/ToggleCommand.java
@@ -1,3 +1,9 @@
+package astraea.command;
+
+import astraea.storage.Storage;
+import astraea.task.TaskList;
+import astraea.ui.AstraeaUI;
+
 import java.io.IOException;
 
 public class ToggleCommand extends Command {
