+++ src/main/java/astraea/storage/Storage.java
@@ -1,3 +1,13 @@
+package astraea.storage;
+
+import astraea.exception.AstraeaFileException;
+import astraea.task.Deadline;
+import astraea.task.Event;
+import astraea.task.Task;
+import astraea.task.TaskList;
+import astraea.task.Todo;
+import astraea.ui.AstraeaUI;
+
 import java.io.BufferedReader;
 import java.io.BufferedWriter;
 import java.io.File;
