+++ src/main/java/astraea/task/Deadline.java
@@ -1,3 +1,7 @@
+package astraea.task;
+
+import astraea.parser.DateParser;
+
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
