+++ src/main/java/astraea/task/Todo.java
@@ -1,3 +1,5 @@
+package astraea.task;
+
 public class Todo extends Task {
     public Todo(String taskName) {
         super(taskName);
